feat(outbound/cache): implement Redis-based RouteCache using bb8-redis with JSON serialization
- Replace stub RouteCache adapter with RedisRouteCache using bb8-redis pool.
- Implement JSON serialization/deserialization of cached plan payloads with serde_json.
- Add real redis-server test harness spawning ephemeral Redis instances for adapter contract tests.
- Move Redis test server harness to shared test_support module for reuse in unit and integration tests.
- Mark integration tests requiring redis-server as ignored by default with instructions for opt-in.
- Update Cargo.toml with required tokio features and optional test-support dependencies.
- Update documentation marking roadmap items 5.1.1 and 5.1.2 completed.

This completes the core Redis cache adapter implementation.
