+++ backend/src/outbound/cache/mod.rs
@@ -5,4 +5,8 @@
 
 mod redis_route_cache;
 
-pub use redis_route_cache::{RedisPool, RedisRouteCache};
+pub use redis_route_cache::RedisRouteCache;
+
+// Export RedisPool only for test-support feature to avoid exposing bb8-redis in public API
+#[cfg(feature = "test-support")]
+pub use redis_route_cache::RedisPool;
