+++ backend/tests/route_cache_redis_bdd.rs
@@ -17,11 +17,12 @@ use rstest::fixture;
 use rstest_bdd::Slot;
 use rstest_bdd_macros::{ScenarioState, given, scenario, then, when};
 use serde::{Deserialize, Serialize};
-use tokio::{net::TcpListener, runtime::Runtime};
+use tokio::runtime::Runtime;
 
 #[path = "support/redis.rs"]
 mod redis_support;
 
+use backend::test_support::redis::unused_redis_url;
 use redis_support::RedisTestServer;
 
 #[derive(Clone)]
@@ -132,15 +133,6 @@ impl RouteCacheWorld {
             .clone();
         self.latest_error.set(error);
     }
-}
-
-async fn unused_redis_url() -> String {
-    let listener = TcpListener::bind("127.0.0.1:0")
-        .await
-        .expect("bind ephemeral port");
-    let address = listener.local_addr().expect("ephemeral port address");
-    drop(listener);
-    format!("redis://{address}/")
 }
 
 #[fixture]
